backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-dotenv
python-jose[cryptography]
httpx
torch
torchaudio
torchvision
audiocraft
demucs
librosa
numpy<2
pydantic>=2.11.0
python-multipart
essentia
supabase
